fix(Assignment): Allows Transnition from failed_to_duplicate to unpublished
This resolves the issue where the duplication is marked as timedout
on the canvas side but the external tool is still processing data.

This allows for a state recovery to occure instead of the user
needing to click the "retry" button manually.

Test Plan:
 - Specs Pass

Fixes: QUIZ-7887
flag = none

Change-Id: Ie6d25510d112a3557e6db75966dc7a8560982987
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/249665
Reviewed-by: Stephen Kacsmark <skacsmark@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Mark McDermott <mmcdermott@instructure.com>
Product-Review: Alex Slaughter <aslaughter@instructure.com>