properly handle german genitive case on profile settings page
closes FOO-363

note that this commit is supposed to be relatively targeted -
there are other uses of genitives (possessives) in Canvas, but I'm
only doing the one at this time, and only writing more complex rules
for german (and even then, only the one case requested in this ticket).
It's already going to require the string it's placed in to be
re-translated in every language (though for many it won't change).
I want to limit exposure until translators have a working example
of how to form their translations when a noun is already in
the genitive case.

test plan:
 * run with RAILS_LOAD_ALL_LOCALES=1
 * go to your settings page, and change your name to something that ends in `x`
   (say... Max)
 * change your language to Deutch (german)
 * save and reload
 * it should say "Max' Settings"
 * change your name to something that doesn't in in x, z, or s (say... Cody)
 * save _and_ reload (the javascript won't work right; again, this is a
   targeted change. once it's proven, we'll make it more robust)
 * it should now say "Codys Settings"
 * change language back to English, save and reload
 * it should now say "Cody's Settings"

Change-Id: Ief17484c5fd6eb9968a1c74102c06ea4891e5108
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252002
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>