Add search to All files
closes LS-1519
flag=rce_enhancements

Back when I added sortBy I chose to bypass putting it in the redux store
for passing the current value down when calling the functions that
eventually make the ajax calls. When adding search, I followed the same
pattern.  It started to become apparent that this was a mistake when
adding search to the Links pane in the tray, but I stuck with it.  Adding
search to the the All Files FileBrowser was the straw that broke it.
In hindsight, my original decision was wrong, colored by my dislike for
redux.  This commit, in addition to adding search to All Files, moves
sort and search into the redux store.  Though the commit looks huge,
it's not as bad as it looks as many of the changes are the same across
slices thru the code for documents, images, media, and links.

The end result is code that's cleaner and more robust in the face of
sorting and searching while the user switches among tray panes, opens
different Link accordion sections, or opens and closes folders in the
file browser.

I'll be glad to walk whoever is CRing this through the changes.

test plan:
  - open all files
  - enter something in the search box
  > expect the "x" clear button to appear
  > expect the files to match
  - open some other folders
  > expect the files in the newly opened folders to match
  - change the search string
  > expect the files in all the open folders to match
  - click the "x" button
  > expect the search field to be cleared
  > expect the the clear  button to disappear
  > expect all the files to be listed

  - enter something in the search box
  - navigate between panels (Documents, Images, Media, Links, All Files),
    contexts (User, Course, Group), and Link accordian sections
  > expect the results to match the current search string

  - it would help if you set DEFAULT_CANVAS_PER_PAGE_ARG in
    canvas-rce-api/app/api/wrapCanvas.js to something
    smaller than 50
  > expect the "Load More" button should load more matching the
    search string
  > expect it not to duplicate the first batch of results

  - open any of the panels
  > while data is loading
    > expect the search input text box to be readonly
      (not disabled, because then it will lose focus)
    > expect the message to change from a search hint to
      a loading message

Change-Id: Ief6cb9f751f0b519efa9a078248e12cf1fc836f9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252389
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>