fix media folder migration
this was hitting a unique constraint violation because it was also
trying to update deleted folders to be hidden. Also running this on
beta took a couple of minutes, so decreased batch size and now it is
a few seconds per batch.

test plan
 - migration should run

fixes VICE-979
flag = none

Change-Id: Icd892ec2e1e8fd1742b9f1916071cad0cd4417ba
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252643
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Caleb Guanzon <cguanzon@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Caleb Guanzon <cguanzon@instructure.com>