Do a more thorough job of extracting LaTeX from math images, etc.
closes LS-1628
flag=new_math_equation_handling

This change fixes the issue found in LS-1627:
The previous code assumed image equations have the data-equation-content
attribute. There are legacy images where that's not true. We now
look at the image's src if it's missing, which is what defines the image
anyway. (I'd look solely at the src, but that makes testing difficult)

While working on 1627, I discovered and fixed a couple bugs around
slowly loading images. First, the `else` was in the wrong place
if the image was not currently loaded, so it would never attach
the onload handler. Also, img.complete is not adequaate to determine
if an image is loaded. This is a bit tricky since complete can be set
if the image isn't really loaded and you can't simply add an onload
handler to all images since it doesn't fire if the image has
already loaded. What does work is to check img.naturalWidth since the
browser can't know that until it's really processed the image data.

test plan:
  - in the rce, create 3 images with the equation LaTeX in the
    data-equation-content, title, or encodeURIComponent twice
    in the src, like this:

<p>data-equation-content:</p>
<p>
  <img class="equation_image"
    title="y = mx + b"
    src="http://localhost:3000/equation_images/y%2520%253D%2520mx%2520%252B%2520b"
    alt="LaTeX: y = mx + b"
    data-equation-content="y = mx + b"
   />
</p>
<p>title:</p>
<p>
  <img class="equation_image"
    title="y = mx + b"
    src="http://localhost:3000/equation_images/y%2520%253D%2520mx%2520%252B%2520b"
    alt="LaTeX: y = mx + b"
  />
</p>
<p>src:</p>
<p>
  <img class="equation_image"
    src="http://localhost:3000/equation_images/y%2520%253D%2520mx%2520%252B%2520b"
    alt="LaTeX: y = mx + b"
  />
</p>
<p>slow:</p>
<p>
  <img class="equation_image"
    src="https://deelay.me/2000/https://latex.codecogs.com/gif.latex?y%20%3D%20x%5E2"
    alt="LaTeX: y = x^2"
    data-equation-content="y = x^2"
  />
</p>

  - save

  - with the new_math_equation_handling (Updated math equation handling)
    flag off
    - refresh the page
    > expect the equation images
 - with the flag on
   - clear your cache (or the delayed image won't be delayed) and refresh the page
   > expect all 4 equations to be replaced with the MathJax version

Change-Id: Ic46ae98447fa5c0eff2d1758ecde7f2fbdaeabe2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/253181
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>