Strip alt attribute from unavailable images in user content
closes LS-1590
flag=none

Ideally we'd strip the alt text in the canvas backend, but that turns
out to be very inconvenient. Instead, add a hidden=1 param to the URL
and strip the alt text in the browser. This leaks the alt text to the
crafty user, but is better than nothing.

If this isn't good enough, I can work through the issues for getting
it handled in the backend.

test plan:
  - as a teacher, unpublish an image file that's embedded in
    RCE content.
  - as a student, view the page
  > expect the lock image in place of the real deal
  > expect the alt text to say the image is unavailable

  - as a teacher, schedule student availability so it is not
    currently available
  > as a student, view the page
  > expect the lock image
  > expect the alt text to say the image is unavailable

Change-Id: Id4c7009bb21cc18227d50ce514159fbf73ff2d15
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/253568
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>