Refine when RCE toolbar buttons are disabled
We disable the content insertion (links, media, etc) toolbar and
menubar buttons in the rce when content like an iframe or video is
selected. If an unlinkable node is selected and the insertion toolbar
buttons are hidden in the overflow menu (i.e. because the width of
the editor is small), the buttons won't be disabled when the overflow
menu is opened. This change fixes that issue.

Also fixes an issue where the disabled content insertion toolbar
buttons that appear to be disabled are still clickable.

fixes LS-1655
flag = rce_enhancements

Test plan:
 - Create a page with an iframe and text on it. Open the RCE in a
   wide window
 - Select the text
 - Expect the content inserton buttons to be enabled and clickable
 - Select the iframe
 - Expect the content insertion buttons to be disabled and not
   clickable
 - Shrink the browser width so that the content insertion buttons
   are hidden
 - Without selecting anything in the editor, click the toolbar
   'more' button
 - Expect the content insertion buttons to be disabled and not
   clickable

Change-Id: If2a2f7cff8047649f027a7b3a6c663c5a896059f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/253910
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>