Fix indent/outdent button when in overflow menu
The indent/outdent button is actually 2 buttons - one regular and
one split. One of the two is always hidden. The hiding and showing
is controlled by a tinymce.dom.DomQuery (tiny's jquery mimic class).
This query is scoped by default to the editor container, but the
overflow menu is not inside the container, so we couldn't show/hide
the buttons if they fell inside the overflow. This commit changes
the scope of the DomQuery to the entire document so we can show/hide
buttons in the overflow.

fixes LS-1593
flag = rce_enhancements

Test plan:
 - In the rce, add some content
 - Click indent and outdent
 - Expect both to work
 - Shrink the window to cause the indent/outdent button to fall into
   overflow menu
 - Click indent and outdent
 - Expect both to work

Change-Id: I6425ea23d61beca988f8771aaeacab5977d5a214
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/254004
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>