Close content tray when option trays are opened
This change updates the link/image/video option trays to explicitly
close any open content tray(s) when they are opened. Prior to this
change the trays relied solely on editor focus/blur triggers to
properly hide themselves when another tray opens, which is a bit
unreliable.

Also, the functionality to return focus back to the editor when a tray
is closed has been made conditional on the active element being
present within the closing tray. This fixes an issue when
showing/hiding multiple trays at once where focus is returned to the
editor when it should really be placed in the newly-opened tray. The
hook to do the focus returning needed to be moved from `onClose` to
`onDismiss` so that we still have access to the `trayRef` for the new
conditional logic.

Finally, this change also adds an event handler to propagate click
events from the editor to the parent document body-- this is necessary
to make instui focus-trapping work properly, since they rely on click
detection to tell when the user is triggering focus via the mouse
intead of the keyboard.

fixes LS-1608
flag = rce_enhancements

Test plan:
  - Create a new page or edit an existing one
  - Add some plain text and a course link (e.g. to an assignment)
  - While focused on the plain text, click on "Insert > Link > Course
    Links" to open the "Add Content" tray again
  - Click on the course link you added in step 2 and select "Link
    Options"
  - Verify that the "Add" tray closes when the "Link Options" tray
    opens
  - You may need to repeat the previous 2 steps a few times to ensure
    that this is fixed, since the buggy behavior was a bit sporadic
  - Do the same with the "Image Options" and "Video Options" trays
  - If you are having trouble reproducing, refer to the video on the
    associated issue

Change-Id: I992a2dfca8937a02a2e9b5f3d74622dbd8795761
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/254808
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Jeff Largent <jeff.largent@instructure.com>