do not sanitize submission draft url
I believe this was erroneously introduced with VICE-91.
We do not dangerously render the url field as html anywhere,
and so it should not be HTML sanitized, and is not a XSS vuln.

fixes DEMO-49
flag=assignments_2_student

test plan:
 - with the FF enabled
 - create an assignment which accepts URL submissions
 - as a student submitting to the assignemnt, type or
   paste a URL containing an ampersand (&)
 - at no point should the student submission be altered
   in any way
 - verify that the user input to the url field cannot be
   maliciously used for XSS

Change-Id: I9d08b04a2f2847d43556e234f991623fac0d0043
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/254838
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Mark McDermott <mmcdermott@instructure.com>
Product-Review: Jared Crystal <jcrystal@instructure.com>
Reviewed-by: Stephen Kacsmark <skacsmark@instructure.com>