create user observer link when enrolling observer
When observers were being added to students via the course roster "Link
to Students" UI the UserObserverLink was not being created. This was
was causing the observer enrollment to be able to get out of sync with
the student they were observing which could cause some view/permissions
bugs. Now the link is being made and the observer's enrollment should
always align with that of the observee's.

fixes EVAL-1296
flag=none

Test Plan:
- Have a course with at least one student and a observer who is currently
 not observing anyone.
- Use the course roster/people view to navigate to the "Link to Students"
view for the observer and start observing a student
- Check the student and observer enrollment and ensure the sections match
- Change the student's section enrollment either changing sections or
adding a section enrollment
- Check the observer enrollment(s) again to see that their section
enrollment matches the student they are observing
- Add the student to another course
- Check to see the observer was added as well and matches the student's
section enrollment
- Remove the student from the course
- Ensure the observer was also removed

Change-Id: I07ac2d6f9760df0488f6263ba01a13ebb1a519b5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/256233
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Syed Hussain <shussain@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>