filters out future sections in observer grades
Observers were able to filter by students who were enrolled in a future
section causing a error to be thrown in some cases. Now students in
future sections with the 'Students can only participate in the course
between these dates' will not be an option in the student drop down or
in the case of that being the only student being observed the observer
would not be given access and shown the 'Not yet available' screen.

fixes EVAL-1404
flag=none

Test Plan:
1. Have a course with 2 students enrolled and another course with 1 of
  those students also enrolled
2. Create 2 sections in the course with both aforementioned students
    - 1 current section with dates including the present day
    - 1 future section with a start date in the future
    - both with the 'Students can only participate in the course
      between these dates' selected
3. Add the student in both courses to the future section and the other
  to the current section
4. Create an observer and link them to both students
5. As the observer navigate to the grades page of the course with both
  students
6. Ensure that only the student in the current section is shown
7. Unselect the 'Students can only participate in the course
    between these dates' in the future section
8. Check to ensure both students are shown on that grade page now

Change-Id: Ia2f7ca68bd93ac5e6780ef8708727ea3372aecc7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/256582
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Syed Hussain <shussain@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>