Fix rce and planner's handling of Persian (fa)
closes LS-1819
flag=none

For some reason, fa is being renamed to fa.json translation file
is being renamed to fa_IR.json (see
https://livegrep.inseng.net/view/transifreq/Jenkinsfile.canvas-lms_community#L23) though when the user selects Persian as their language in canvas, our
lang is set to 'fa'. This caused planner to render in English and
the RCE to blow up.

With this change both planner and the rce will load the fa_IR file
when the language is set to 'fa'. Additionally I made a little
change in the RCE so it will always fall back to English if it
can't load the translations for the current language.

Ideally I'd like to stop renaming it, then mapping fa-* to 'fa', but
this change would still have to happen until translations made another
round-trip.

test plan:
  - build canvas with all locals included:
    bundle exec rake i18n:generate_js
    yarn build:packages
    RAILS_LOAD_ALL_LOCALES=1 yarn run webpack
  - in /profile/settings, select another language
  > expect the student planner and the new RCE to be in
    the language (for the RCE you can look at tooltips on
    the Bold, Italic etc. toolbar buttons.
  - change the language to Persian (in the dropdown list,
    it's 5th from the bottom)
  > expect the planner and RCE to be in Persian

Change-Id: I85cbfddc1a4dad568e3ae28dc8760be1dc1cf280
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/257231
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>