show correct reason for date locked assignments being unavailable
When an assignment is date locked and in a future module, it displays as
locked due to prerequisites instead and links to the previous assignment
in module order. It should instead display that the assignment is locked
until a specific date.

closes EVAL-1416
flag=assignments_2_student

Test Plan:
- Have course with at least one student
- Create an assignment that is not yet available until a future date
- Create two modules
- Put the assignment just created in the second module and anything else
in the first module besides the title
- Navigate to the assignment created as the student and make sure the
correct date is displayed as to when the assignment will be available
- Ensure it doesn't mention or link to any prerequisites

Change-Id: Ib4a4d4a977d38135838c44ff97adfce650b175fa
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/257710
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Adrian Packel <apackel@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Gary Mei <gmei@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>