Allow setting 0 on middle tier of conditional release
Previously, 0 could be set on middle tier of conditional release,
but weird behavior would manifest when saving, including changing
the score to the max score possible, or leaving the input blank.
This was due to 2 separate issues:

In value(), the isUpperBound flag passed to transformScore() should
not be set - the changeable boundaries in the cyoe editor both
represent lower bounds, not upper bounds.

In the reducer, the identity() function would return the empty string
(default) if the state value was 0, since the function was checking
if the value was falsey (i.e. 0). Instead, we should return dflt if
the value is nullish, but not if its 0.

flag=conditional_release
fixes LS-1799

Test plan:
 - Create a quiz with a few questions
 - Create a few pages, set pages to be allowed in mastery paths
 - Edit the quiz, choose mastery paths tab
 - Enter a valid score in the first tier
 - In the second tier, enter 0
 - Save the quiz
 - Open mastery paths again, expect the 0 to still be there
 - Repeat once more, expect the 0 to persist in the UI
 - Mastery path should give correct page to a student that takes the
   quiz

Change-Id: I7b7b082a2760ae7b4472c3cb4094d7cf5dc61928
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/257770
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>